test(sessioncatalog): keep path migration regression within lint limits
Move the v9 projection invalidation regression into its own focused test file so the existing catalog test file remains below the repository 800-line ceiling.

Verification:
- go test ./internal/sessioncatalog
- go run ./tools/repolint
